Update to S3Boto3Storage
The current file storage backend, S3BotoStorage, cannot be used for
Python versions >=3.4.4 (https://github.com/boto/boto/issues/3433).
Therfore we need to update the plugin to use the S3Boto3Storage which
uses a different configuration logic. As a result, the following
configuration parameters are changed:
- S3_ADDRESSING_STYLE is added;
- S3_SIGNATURE_VERSION is added;
- S3_REGION is added;
- S3_AUTO_CREATE_BUCKET is removed;

Also we switch to presigned URLs which removes the need for
setting bucket ACL to be public.

Additionally, "contrib" is added to the package name to differentiate
this plugin from official Tutor plugins.

Fixes: #8