docs: Update README to mention pip install from GitHub
As long as we don't have a release up on PyPI, instruct people to
install from Git, pointing them to the most recent release.

Update .bumpversion.cfg so that we update the README with the latest
version on every release.

Also, fix a double colon ("::") in the README that was left over from
the ReST→Markdown conversion.