docs: Remove wrapped paragraphs in bug report template
When rendered in GitHub for opening an issue, the 80-column line wrap
looks bad in the preview. Since everything about the bug report
template is for making it easy to submit good bug reports, and UI
impressions matter for perceived ease of use, remove the extraneous
newlines and let the GitHub issue UI reflow the text as needed.