fix: Correct data type for defaults
Jinja2 templates apparently don't like when a variable has the value
None. Although they behave correctly when the variable is used in an
{% if VARIABLE %} statement (where None evaluates to a falsy value),
in a substitution ({{ VARIABLE }}) they substitute the string "none"
rather than he empty string.

That's unhelpful, thus replace the None defaults with the empty
string. This way, they still evaluate to a falsy value in an {% if %}
statement, but are also substituted as the empty string.

Also, fix the "S3_USE_SSL" default to a boolean True, rather than the
string "True".