feat: add ability to disable profile pictures to use S3 bucket
Without this patch when the plugin is activated all file are saved in an
S3 Bucket or an API-compatible storage. We don't have the option to keep
the profile pictures in the filesystem.

This is a problem because Open edX makes it possible to configure
individually each type of data, such as Grades, Reports, Profiles
Pictures, ORA2 files, etc.

This patch solves the problem by add the ability to continue using the
filesystem for profile pictures when the plugin is activated.

Use:
- When `S3_PROFILE_IMAGE_BUCKET` is empty the profile pictures will be
  saved in filesystem. The default `PROFILE_IMAGE_BACKEND` won't be
  overwritten.

- When `S3_PROFILE_IMAGE_BUCKET` is set the profile pictures will
  be saved in an S3 Bucket that must have a public ACL. The
  `PROFILE_IMAGE_BACKEND` will be overwritten.

Resolves: #40