build: add python virtual environment as a flake8 exclusion directory
Without this patch, we need to execute `commit` and `push` with
`--no-verify` because flake8 is not bypassing the python virtual
environment and it causes errors. We also are unable to execute a
`rebase` without deleting tox files before.

This patch solves the problem by adding the `venv*` folder in flake8
exclude property and also add generated tox files in `.gitignore` file.