docs: Update version compatibility matrix in README
We started supporting Nutmeg with 1.0.0, Olive with 1.1.0, and Palm
with 1.2.0. Technically all three Open edX releases *should* work in
conjunction with whatever is the latest state of `main`, but it
probably doesn't hurt to recommend that people use a minor version to
match the Open edX release which we regularly test at the time we tag
the minor.

Thus, update the version compatibility matrix in the README
accordingly.